admin: factor out config routes
